Store intermediate info and final stats in CSV
This commit refactors the import/export stat scripts so that
intermediate results like the list of PRs and import/export-PR mapping
are persisted in CSV files sorted by a primary key (PR number), which
enables us to check in these CSV files with clean and meaningful history
(new information is expected to be appended to the files). Benefits
include:
* incrementally processing new PRs and calculating mappings (faster!);
* easily producing histograms of import/export latencies.

(csv_database.py is a simple line-based, single-primary-key, CSV-backed
"database" for this purpose.)

Lastly, stats are changed to weekly basis instead of monthly.